[MOD] XQuery: refactorings (functions classes renamed) [ADD] XQuery: some code eliminations and rewritings added

+++ src/main/java/org/basex/util/Token.java
@@ -290,10 +290,10 @@ public final class Token {
    * @param token token
    * @return codepoints
    */
-  public static int[] cps(final byte[] token) {
+  public static long[] cps(final byte[] token) {
     int pos = 0;
     final int len = token.length;
-    final int[] cp = new int[len];
+    final long[] cp = new long[len];
     for(int i = 0; i < len; i += cl(token, i)) cp[pos++] = cp(token, i);
     return pos < len ? Arrays.copyOf(cp, pos) : cp;
   }
